Begin splitting up JSON serializer from core reflect code
