+++ source/library/reflect/public/potato/reflect/CMakeLists.txt
@@ -3,5 +3,6 @@ target_sources(reflect PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/_wrapper.h
     ${CMAKE_CURRENT_SOURCE_DIR}/metadata.h
     ${CMAKE_CURRENT_SOURCE_DIR}/reflect.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/json_serializer.h
     ${CMAKE_CURRENT_SOURCE_DIR}/stream_serializer.h
 )
