+++ source/library/reflect/tests/test_serialize.cpp
@@ -1,5 +1,6 @@
 #include "potato/reflect/reflect.h"
 #include "potato/reflect/stream_serializer.h"
+#include "potato/reflect/json_serializer.h"
 #include <doctest/doctest.h>
 #include <sstream>
